Move spline to a common file
