+++ parse_pwe_templates.py
@@ -1181,76 +1181,6 @@ class AnimationParameters(BaseProperty):
 """
     )
     modules.append("AnimationParameters")
-    core_path.joinpath("Spline.py").write_text(
-        """# Generated file
-import dataclasses
-import struct
-import typing
-import typing_extensions
-
-import construct
-
-from retro_data_structures.common_types import MayaSpline
-from retro_data_structures.game_check import Game
-from retro_data_structures.properties.base_spline import BaseSpline, Knot
-
-
-def _read_knot(data: typing.BinaryIO) -> Knot:
-    header = typing.cast(tuple[float, float, int, int], struct.unpack(">ffBB", data.read(10)))
-    cached_tangents_a = None
-    cached_tangents_b = None
-    if header[2] == 5:
-        cached_tangents_a = typing.cast(tuple[float, float], struct.unpack(">ff", data.read(8)))
-    if header[3] == 5:
-        cached_tangents_b = typing.cast(tuple[float, float], struct.unpack(">ff", data.read(8)))
-
-    return Knot(*header, cached_tangents_a=cached_tangents_a, cached_tangents_b=cached_tangents_b)
-
-
-@dataclasses.dataclass()
-class Spline(BaseSpline):
-
-    @classmethod
-    def from_stream(cls, data: typing.BinaryIO, game: Game, size: int | None = None) -> typing_extensions.Self:
-        pre_infinity, post_infinity, knot_count = struct.unpack(">BBL", data.read(6))
-        knots = [
-            _read_knot(data)
-            for _ in range(knot_count)
-        ]
-        clamp_mode, minimum_amplitude, maximum_amplitude = struct.unpack(">Bff", data.read(9))
-
-        return cls(
-            pre_infinity=pre_infinity,
-            post_infinity=post_infinity,
-            knots=knots,
-            clamp_mode=clamp_mode,
-            minimum_amplitude=minimum_amplitude,
-            maximum_amplitude=maximum_amplitude,
-        )
-
-    def to_stream(self, data: typing.BinaryIO, game: Game) -> None:
-        MayaSpline.build_stream(construct.Container(
-            pre_infinity=self.pre_infinity,
-            post_infinity=self.post_infinity,
-            knots=[
-                construct.Container(
-                    time=knot.time,
-                    amplitude=knot.amplitude,
-                    unk_a=knot.unk_a,
-                    unk_b=knot.unk_b,
-                    cached_tangents_a=knot.cached_tangents_a,
-                    cached_tangents_b=knot.cached_tangents_b,
-                )
-                for knot in self.knots
-            ],
-            clamp_mode=self.clamp_mode,
-            minimum_amplitude=self.minimum_amplitude,
-            maximum_amplitude=self.maximum_amplitude,
-        ), data)
-
-"""
-    )
-    modules.append("Spline")
 
     create_all_file(core_path.joinpath("__init__.py"), base_import, modules)
 
@@ -1527,7 +1457,12 @@ def parse_game(templates_path: Path, game_xml: Path, game_id: str) -> dict:
                 dependency_code = "{obj}.dependencies_for(asset_manager)"
             else:
                 prop_type = raw_type
-            needed_imports[f"{import_base}.core.{prop_type}"] = prop_type
+
+            if raw_type == "Spline":
+                needed_imports[f"retro_data_structures.properties.{raw_type.lower()}"] = prop_type
+            else:
+                needed_imports[f"{import_base}.core.{prop_type}"] = prop_type
+
             parse_code = f"{prop_type}.from_stream(data, game, property_size)"
             build_code.append("{obj}.to_stream(data, game)")
             from_json_code = f"{prop_type}.from_json({{obj}})"
