+++ src/retro_data_structures/properties/spline.py
@@ -1,8 +1,12 @@
 from __future__ import annotations
 
 import dataclasses
+import struct
 import typing
 
+import construct
+
+from retro_data_structures.common_types import MayaSpline
 from retro_data_structures.properties.base_property import BaseProperty
 from retro_data_structures.properties.field_reflection import FieldReflection
 
@@ -10,6 +14,7 @@ if typing.TYPE_CHECKING:
     import typing_extensions
 
     from retro_data_structures import json_util
+    from retro_data_structures.game_check import Game
 
     class KnotJson(typing_extensions.TypedDict):
         time: float
@@ -115,8 +120,20 @@ class Knot(BaseProperty):
         }
 
 
+def _read_knot(data: typing.BinaryIO) -> Knot:
+    header = typing.cast("tuple[float, float, int, int]", struct.unpack(">ffBB", data.read(10)))
+    cached_tangents_a = None
+    cached_tangents_b = None
+    if header[2] == 5:
+        cached_tangents_a = typing.cast("tuple[float, float]", struct.unpack(">ff", data.read(8)))
+    if header[3] == 5:
+        cached_tangents_b = typing.cast("tuple[float, float]", struct.unpack(">ff", data.read(8)))
+
+    return Knot(*header, cached_tangents_a=cached_tangents_a, cached_tangents_b=cached_tangents_b)
+
+
 @dataclasses.dataclass()
-class BaseSpline(BaseProperty):
+class Spline(BaseProperty):
     pre_infinity: int = dataclasses.field(
         default=0,
         metadata={
@@ -153,6 +170,44 @@ class BaseSpline(BaseProperty):
             "reflection": FieldReflection[float](float, id=0x00000005, original_name="MaximumAmplitude"),
         },
     )
+
+    @classmethod
+    def from_stream(cls, data: typing.BinaryIO, game: Game, size: int | None = None) -> typing_extensions.Self:
+        pre_infinity, post_infinity, knot_count = struct.unpack(">BBL", data.read(6))
+        knots = [_read_knot(data) for _ in range(knot_count)]
+        clamp_mode, minimum_amplitude, maximum_amplitude = struct.unpack(">Bff", data.read(9))
+
+        return cls(
+            pre_infinity=pre_infinity,
+            post_infinity=post_infinity,
+            knots=knots,
+            clamp_mode=clamp_mode,
+            minimum_amplitude=minimum_amplitude,
+            maximum_amplitude=maximum_amplitude,
+        )
+
+    def to_stream(self, data: typing.BinaryIO, game: Game) -> None:
+        MayaSpline.build_stream(
+            construct.Container(
+                pre_infinity=self.pre_infinity,
+                post_infinity=self.post_infinity,
+                knots=[
+                    construct.Container(
+                        time=knot.time,
+                        amplitude=knot.amplitude,
+                        unk_a=knot.unk_a,
+                        unk_b=knot.unk_b,
+                        cached_tangents_a=knot.cached_tangents_a,
+                        cached_tangents_b=knot.cached_tangents_b,
+                    )
+                    for knot in self.knots
+                ],
+                clamp_mode=self.clamp_mode,
+                minimum_amplitude=self.minimum_amplitude,
+                maximum_amplitude=self.maximum_amplitude,
+            ),
+            data,
+        )
 
     @classmethod
     def from_json(cls, data: json_util.JsonValue) -> typing_extensions.Self:
